refactor: 차트 추가 다이얼로그들
