+++ components/hospital/icu/main/chart/chart-entry.tsx
@@ -1,10 +1,10 @@
 'use client'
 
 import NoResultSquirrel from '@/components/common/no-result-squirrel'
-import AddChartDialogs from '@/components/hospital/icu/main/chart/add-chart-dialogs/add-chart-dialogs'
 import Chart from '@/components/hospital/icu/main/chart/chart'
 import { useBasicHosDataContext } from '@/providers/basic-hos-data-context-provider'
 import type { SelectedChart } from '@/types/icu/chart'
+import PasteChartDialogs from './paste-chart-dialogs/paste-chart-dialogs'
 
 export default function ChartEntry({
   chartData,
@@ -36,13 +36,17 @@ export default function ChartEntry({
 
   // io가 있고 chart가 없음 => 첫날 차트가 아님
   if (hasIcuIo && !chartData) {
-    return <AddChartDialogs chartData={chartData} />
+    return <PasteChartDialogs chartData={chartData} />
   }
 
   // io가 있고 chart가 있고 order가 없는 경우 => 첫날차트
   if (hasIcuIo && chartData && chartData.orders.length === 0) {
     return (
-      <AddChartDialogs chartData={chartData} patientId={patientId} firstChart />
+      <PasteChartDialogs
+        chartData={chartData}
+        patientId={patientId}
+        firstChart
+      />
     )
   }
